backend/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
flask-openapi3==2.3.2
environs==9.5.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
eventlet==0.30.2
pydantic==1.10.8
pytest==7.3.1
freezegun==1.2.2
sqlalchemy-utils==0.37.8